+++ src/keys.rs
@@ -40,6 +40,7 @@ const_key!(HOME, Home);
 const_key!(LEFT, Left);
 const_key!(QUESTION, Char('?'));
 const_key!(RIGHT, Right);
+const_key!(SPACE, Char(' '));
 const_key!(TAB, Tab);
 const_key!(UP, Up);
 
@@ -126,6 +127,7 @@ pub fn parse_key(raw: &str) -> Result<KeyEvent, ConfError> {
         "f10" => F(10),
         "f11" => F(11),
         "f12" => F(12),
+        "space" => Char(' '),
         c if c.len() == 1 => Char(c.chars().next().unwrap()),
         _ => {
             return bad_key(raw);